requirements.txt
# Core geospatial libraries
rasterio>=1.3.0
geopandas>=0.14.0
shapely>=2.0.0
pyproj>=3.6.0
fiona>=1.9.0

# Image processing and computer vision
opencv-python>=4.8.0
scikit-image>=0.22.0
numpy>=1.24.0
scipy>=1.11.0
pillow>=10.0.0

# Data manipulation and analysis
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# 2D Visualization
folium>=0.14.0
plotly>=5.17.0

# 3D Visualization
pyvista>=0.42.0
pyvistaqt>=0.11.0

# Report generation
reportlab>=4.0.0
jinja2>=3.1.0
weasyprint>=60.0

# GUI
streamlit>=1.28.0
tkinter-tooltip>=2.1.0

# Additional utilities
tqdm>=4.66.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Web API
flask>=3.0.0
flask-cors>=4.0.0
